--- a/tokio/src/util/mod.rs
+++ b/tokio/src/util/mod.rs
@@ -81,3 +81,5 @@ pub(crate) mod error;
 pub(crate) mod memchr;
 
 pub(crate) mod markers;
+
+pub(crate) mod cacheline;