@carllerche 2023-06-28
The unfortunate issue with this change is it makes creating a channel much more expensive for "one shot" operations.