@spongecaptain 2023-06-29
> What system are you running the benchmark on?

On both Linux and Darwin.