@spongecaptain 2023-06-29
> What system are you running the benchmark on?

The benchmark difference is almost the same on both systems.