@spongecaptain 2023-06-29
> The unfortunate issue with this change is it makes creating a channel much more expensive for "one shot" operations.

I known, but this may not the first of all issue. Because of zero cost in rust, in oneshot scene users may use oneshot instead. The external cost may accepted, though I would like to found how it happen.