@spongecaptain 2023-06-29
> Could you experiment w/ only padding 64 instead of 128. Does it make a difference?

Which size to pad, it depends on the which cpu you run. The size of cache line is designed by hardware.