@spongecaptain 2023-06-29
The reason why create after using `CachePadded` might be using extra function call `memcpy`.

The source code:

```rust
fn get_s() -> Box<S> {
    Box::new(S { i: 0, j: 0 })
}

struct S {
    i: usize,
    j: usize,
}


#[repr(align(128))]
struct Align{
    i:usize,
    j:usize,
}

fn get_align()-> Box::<Align> {
    Box::new(Align { i: 0, j:0 })
}

fn main(){
    let s = get_s();
    let a = get_align();
}
```
The assembly code of two function `get_s()` and `get_align()` is
```
playground::get_s: # @playground::get_s
# %bb.0:
	subq	$72, %rsp
	movq	$0, 24(%rsp)
	movq	$0, 32(%rsp)
	movq	24(%rsp), %rcx
	movq	%rcx, (%rsp)                    # 8-byte Spill
	movq	32(%rsp), %rax
	movq	%rax, 8(%rsp)                   # 8-byte Spill
	movq	%rcx, 40(%rsp)
	movq	%rax, 48(%rsp)
	movl	$16, %edi
	movl	$8, %esi
	callq	alloc::alloc::exchange_malloc
	movq	%rax, 16(%rsp)                  # 8-byte Spill
	jmp	.LBB15_2
	movq	%rax, %rcx
	movl	%edx, %eax
	movq	%rcx, 56(%rsp)
	movl	%eax, 64(%rsp)
	movq	56(%rsp), %rdi
	callq	_Unwind_Resume@PLT
	ud2

.LBB15_2:
	movq	16(%rsp), %rax                  # 8-byte Reload
	movq	8(%rsp), %rcx                   # 8-byte Reload
	movq	(%rsp), %rdx                    # 8-byte Reload
	movq	%rdx, (%rax)
	movq	%rcx, 8(%rax)
	addq	$72, %rsp
	retq
                                        # -- End function

playground::get_align: # @playground::get_align
# %bb.0:
	pushq	%rbp
	movq	%rsp, %rbp
	andq	$-128, %rsp
	subq	$384, %rsp                      # imm = 0x180
	movq	$0, 128(%rsp)
	movq	$0, 136(%rsp)
	movl	$128, %esi
	movq	%rsi, %rdi
	callq	alloc::alloc::exchange_malloc
	movq	%rax, 120(%rsp)                 # 8-byte Spill
	jmp	.LBB16_2
	movq	%rax, %rcx
	movl	%edx, %eax
	movq	%rcx, 352(%rsp)
	movl	%eax, 360(%rsp)
	movq	352(%rsp), %rdi
	callq	_Unwind_Resume@PLT
	ud2

.LBB16_2:
	movq	120(%rsp), %rdi                 # 8-byte Reload
	leaq	128(%rsp), %rsi
	movl	$128, %edx
	callq	memcpy@PLT
	movq	120(%rsp), %rax                 # 8-byte Reload
	movq	%rbp, %rsp
	popq	%rbp
	retq
                                        # -- End function
```

As we can see, the funciton `get_align()` use external `memcpy` to copy 128 bytes from stack to heap, but `get_s()` just use basic instructions like `movq`. This difference may cause the more CPU cost when using `CachePadded`.

You can see more detail about the above assembly code in https://play.rust-lang.org/?version=nightly&mode=debug&edition=2021&gist=7ddf997fbb5e23ceaf4a9ab485f01771