@Darksonn 2023-07-03
I'm not getting a memcpy in the playground when I try it.
```
playground::get_s: # @playground::get_s
# %bb.0:
	pushq	%rax
	movq	__rust_no_alloc_shim_is_unstable@GOTPCREL(%rip), %rax
	movzbl	(%rax), %eax
	movl	$16, %edi
	movl	$8, %esi
	callq	*__rust_alloc@GOTPCREL(%rip)
	testq	%rax, %rax
	je	.LBB0_1
# %bb.2:
	xorps	%xmm0, %xmm0
	movups	%xmm0, (%rax)
	popq	%rcx
	retq

.LBB0_1:
	movl	$8, %edi
	movl	$16, %esi
	callq	*alloc::alloc::handle_alloc_error@GOTPCREL(%rip)
	ud2
                                        # -- End function

playground::get_align2: # @playground::get_align2
# %bb.0:
	pushq	%rax
	movq	__rust_no_alloc_shim_is_unstable@GOTPCREL(%rip), %rax
	movzbl	(%rax), %eax
	movl	$128, %edi
	movl	$128, %esi
	callq	*__rust_alloc@GOTPCREL(%rip)
	testq	%rax, %rax
	je	.LBB1_1
# %bb.2:
	xorps	%xmm0, %xmm0
	movaps	%xmm0, (%rax)
	popq	%rcx
	retq

.LBB1_1:
	movl	$128, %edi
	movl	$128, %esi
	callq	*alloc::alloc::handle_alloc_error@GOTPCREL(%rip)
	ud2
                                        # -- End function
```
[playground](https://play.rust-lang.org/?version=nightly&mode=release&edition=2021&gist=b6ea5d77b19f0f050a81ea54f5dcbff7)