@wathenjiang 2023-07-03
Sorry for that, it is a mistake, we should  assemble the above code in the release mode.
So the reason why creating a aligned struct is less efficient might be the parameter to [`alloc::alloc::__rust_alloc`](https://stdrs.dev/nightly/x86_64-unknown-linux-gnu/alloc/alloc/fn.__rust_alloc.html#) is different, the performance of allocation on heap is affected by `size` and `align`.