@wathenjiang 2023-07-29
> There are some CI failures. Please ensure that CI passes. (And you probably want to merge in master so that you are up-to-date with CI changes.)

The current CI error is
```
Run cargo hack test --each-feature
info: running `cargo test --no-default-features` on tests-integration (1/10)
   Compiling tokio v1.29.1 (/home/runner/work/tokio/tokio/tokio)
error: associated function `new` is never used
  --> tokio/src/util/cacheline.rs:77:19
   |
75 | impl<T> CachePadded<T> {
   | ---------------------- associated function in this implementation
76 |     /// Pads and aligns a value to the length of a cache line.
77 |     pub(crate) fn new(value: T) -> CachePadded<T> {
   |                   ^^^
   |
   = note: `-D dead-code` implied by `-D warnings`

error: could not compile `tokio` (lib) due to previous error
error: process didn't exit successfully: `/home/runner/.rustup/toolchains/stable-x86_64-unknown-linux-gnu/bin/cargo test --manifest-path Cargo.toml --no-default-features` (exit status: 101)
Error: Process completed with exit code 1.
```
And another error is the same.
But we do call the `CachePadded::new` in `mpsc::channel()` method.
Do you have some suggestions to make it pass?