@Darksonn 2023-07-29
If the `sync` feature is disabled, then you're not using `CachePadded`. You'll need to add conditional compilation to remove it in those cases.