@wathenjiang 2023-08-02
The other sync tools in `tokio::sync` may could benefit from applying `CachePadded`. For example, the `tokio::sync::broadcast`.

I initially wrote a benchmark test:https://github.com/wathenjiang/tokio-2.8.2-coderead/blob/bfe26af3b56fe8eac81b6087829eb4c17c88fb49/benches/sync_broadcast.rs#L1-L46

And apply `CachePadded` in https://github.com/wathenjiang/tokio-2.8.2-coderead/blob/bfe26af3b56fe8eac81b6087829eb4c17c88fb49/tokio/src/sync/broadcast.rs#L306-L318

The following is a performance comparison test before and after applying `CachePadded`

**Before**

```
# cargo bench --bench sync_broadcast
    Finished bench [optimized] target(s) in 0.05s
     Running sync_broadcast.rs (/root/rustproject/tokio-2.8.2-coderead/target/release/deps/sync_broadcast-a42a48a6aede9bb7)

running 1 test
test contention_unbounded ... bench:  46,916,459 ns/iter (+/- 5,925,933)

test result: ok. 0 passed; 0 failed; 0 ignored; 1 measured
```

**after**

```
# cargo bench --bench sync_broadcast
    Finished bench [optimized] target(s) in 0.05s
     Running sync_broadcast.rs (/root/rustproject/tokio-2.8.2-coderead/target/release/deps/sync_broadcast-a42a48a6aede9bb7)

running 1 test
test contention_unbounded ... bench:  41,510,877 ns/iter (+/- 7,768,113)

test result: ok. 0 passed; 0 failed; 0 ignored; 1 measured
```

An increase of approximately 11%.